backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
aiohttp==3.8.6
requests==2.31.0
python-multipart==0.0.6
httpx==0.25.1
logging-plus==1.2.0
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.21.1